+++ lib/iris/tests/test_coding_standards.py
@@ -76,9 +76,9 @@ class StandardReportWithExclusions(pep8.StandardReport):
         '*/iris/analysis/trajectory.py',
         '*/iris/fileformats/cf.py',
         '*/iris/fileformats/dot.py',
-        '*/iris/fileformats/grib/__init__.py',
         '*/iris/fileformats/grib/_grib_cf_map.py',
         '*/iris/fileformats/grib/load_rules.py',
+        '*/iris/fileformats/grib/_gribwrapper.py',
         '*/iris/fileformats/pp.py',
         '*/iris/fileformats/pp_rules.py',
         '*/iris/fileformats/rules.py',
